+++ rr/txt.js
@@ -72,9 +72,7 @@ export default class TXT extends RR {
     // test.example.com  3600  IN  TXT  "..."
     const match = opts.bindline
       .trim()
-      .split(
-        /^([\S]{1,255})\s+([0-9]{1,10})\s+(IN)\s+(\w{3})\s+?\s*(.*?)$/i,
-      )
+      .split(/^([\S]{1,255})\s+([0-9]{1,10})\s+(IN)\s+(\w{3})\s+?\s*(.*?)$/i)
     if (!match) this.throwHelp(`unable to parse TXT: ${opts.bindline}`)
     const [owner, ttl, c, type, rdata] = match.slice(1)